+++ internal/taints/taints.go
@@ -141,7 +141,7 @@ func preferNoSchedule(client *kubernetes.Clientset, nodeID, taintName string, ef
 			},
 		}
 	} else {
-		// add missing taint to exsting list
+		// add missing taint to existing list
 		patches = []patchTaints{
 			{
 				Op:    "add",
