+++ internal/timewindow/days.go
