+++ pkg/checkers/commandchecker_test.go
